Add template inspection functions
Added a template inspection function to retrieve all properties
being referenced by a mustache template. The values returned are
objects with get()/set() methods to retrieve and set the property
respectively. Think of these objects as property accessors. This
function is useful when property inspection is needed before rendering
a mustache template.